Tri-Coins
##PROBLEM
Elizabeth belongs to a very rich family which owns many gold mines. Today, she brought N gold coins and decided to form a triangle using these coins. Isn't it strange?

Elizabeth has a unusual way of forming a triangle using gold coins, which is described as follows:

She puts 1 coin in the 1st row. then puts 2 coins in the 2nd row. then puts 3 coins in the 3rd row. and so on as shown in the given figure. .

She is interested in forming a triangle with maximum possible height using at most N coins. Can you tell her the maximum possible height of the triangle?