Height of Rationals
In a recent breakthrough in mathematics, the proof utilized a concept called Height.
Consider a fraction a/b
​
Its Height is defined as the maximum of its numerator and denominator. So, for example, the Height of 19/3 would be 19, and the Height of  4/27 would be 27.

Given a and b, find the Height of a/b.