Dominant Army
In the medieval age, there were 3 kingdoms A, B, and C. The army of these kingdom had N_A , N_B , and N_C soldiers respectively.

You are given that an army with X soldiers can defeat an army with Y soldiers only if X>Y.

An army is said to be dominant if it can defeat both the other armies combined. For example, kingdom C's army will be dominant only if N_C > N_A + N_B
Determine whether any of the armies is dominant or not.