Split and Maximize
Let f be a function, such that, for an array A of size M, f(A) is defined as
f(A) =∑(M;i=1) ∑(M=j;j!=i) (A i ⋅A j)

You are given an array C of size N. In one operation on the array, you can:

Choose an index i (1≤i≤∣C∣)
Select two positive integers X and Y such that X+Y = C_i
​
Remove the element C_i from the array C
Insert the elements X and Y to the array C
Find the maximum value of f(C) that can be obtained using any (possibly zero) number of given operations. Since this value can be huge, output the answer modulo 998244353998244353.