Minimum number of Flips
Chef has an array A of length N consisting of 1 and -1 only.

In one operation, Chef can choose any index i(1≤i≤N) and multiply the element A_i  by −1.

Find the minimum number of operations required to make the sum of the array equal to 0. Output -1 if the sum of the array cannot be made 0.