Degree of Polynomial
In mathematics, the degree of polynomials in one variable is the highest power of the variable in the algebraic expression with non-zero coefficient.

Chef has a polynomial in one variable x with NN terms. The polynomial looks like A_0.x^0 + A_1.x^1 + ... + A_N-2.x^N-2 + A_N-1.x^N-1 where A_i−1 denotes the coefficient of the ith term x^i−1 for all (1≤i≤N).

Find the degree of the polynomial.

Note: It is guaranteed that there exists at least one term with non-zero coefficient.