Equalizing Numbers
Chef has two integers A and B. In one operation he can choose any integer d, and make one of the following two moves :

Add d to A and subtract d from B.
Add d to B and subtract d from A.
Chef is allowed to make as many operations as he wants. Can he make A and B equal?