Good Weather
The weather report of Chefland is Good if the number of sunny days in a week is strictly greater than the number of rainy days.

Given 7 integers A_1, A_2, A_3, A_4, A_5, A_6, A_7 where A_i =1 denotes that the ith day of week in Chefland is a sunny day, A_i = 0 denotes that the ith day in Chefland is a rainy day. Determine if the weather report of Chefland is Good or not.