Sums in a Triangle
Given an integer N, let us consider a triangle of numbers of N lines in which a number a11 appears in the first line, two numbers a21 and a22 appear in the second line, three numbers a31, a32 and a33 appear in the third line, etc. In general, i numbers ai1 ,ai2 …aii appear in the ith line for all 1≤i≤N. Develop a program that will compute the largest of the sums of numbers that appear on the paths starting from the top towards the base, so that:

on each path the next number is located on the row below, more precisely either directly below or below and one place to the right.