Secret Recipe
Chef and his competitor Kefa own two restaurants located at a straight road. The position of Chef's restaurant is X_1, the position of Kefa's restaurant is X_2.
Chef and Kefa found out at the same time that a bottle with a secret recipe is located on the road between their restaurants. The position of the bottle is X_3.
The cooks immediately started to run to the bottle. Chef runs with speed V_1, Kefa with speed V_2.
Your task is to figure out who reaches the bottle first and gets the secret recipe (of course, it is possible that both cooks reach the bottle at the same time).