N Buttons 1 Bulb
Kulyash stays in room that has a single bulb and N buttons. The bulb is initially on.

The initial states of the buttons are stored in a binary string S of length N — if S_i is 0, the i-th button is off, and if S_i is 1, the i-th button is on. If Kulyash toggles any single button then the state of the bulb reverses i.e. the bulb lights up if it was off and vice versa.

Kulyash has toggled some buttons and the final states of the buttons are stored in another binary string R of length N. He asks you to determine the final state of the bulb.